@@ -3,93 +3,74 @@
     <modelVersion>4.0.0</modelVersion>
     <groupId>com.andyinthecloud.githubsfdeploy</groupId>
     <artifactId>githubsfdeploy</artifactId>
-    <packaging>war</packaging>
+    <packaging>jar</packaging>
     <version>0.0.2-SNAPSHOT</version>
-
+    <parent>
+        <groupId>org.springframework.boot</groupId>
+        <artifactId>spring-boot-starter-parent</artifactId>
+        <version>3.4.5</version>
+        <relativePath/>
+    </parent>
+    
     <properties>
-        <org.springframework.version>4.2.5.RELEASE</org.springframework.version>
-        <org.slf4j.version>1.6.2</org.slf4j.version>
-        <com.github.ryanbrainard.richsobjects.version>0.2-BETA</com.github.ryanbrainard.richsobjects.version>
-        <com.force.sdk.version>22.0.9-BETA</com.force.sdk.version>
+        <maven.compiler.release>17</maven.compiler.release>
+        <org.slf4j.version>2.0.13</org.slf4j.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
 
     <dependencies>
         <dependency>
-            <groupId>javax.servlet</groupId>
-            <artifactId>javax.servlet-api</artifactId>
-            <version>3.1.0</version>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-web</artifactId>
+        </dependency>    
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-security</artifactId>
+        </dependency>       
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-tomcat</artifactId>
+            <scope>provided</scope>
+        </dependency>        
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-oauth2-client</artifactId>
+        </dependency>        
+        <dependency>
+            <groupId>jakarta.servlet</groupId>
+            <artifactId>jakarta.servlet-api</artifactId>
             <scope>provided</scope>
         </dependency>
         <dependency>
-            <groupId>org.springframework</groupId>
-            <artifactId>spring-context</artifactId>
-            <version>${org.springframework.version}</version>
-            <exclusions>
-                <!-- Exclude Commons Logging in favor of SLF4j -->
-                <exclusion>
-                    <groupId>commons-logging</groupId>
-                    <artifactId>commons-logging</artifactId>
-                </exclusion>
-            </exclusions>
+            <groupId>jakarta.servlet.jsp.jstl</groupId>
+            <artifactId>jakarta.servlet.jsp.jstl-api</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.springframework</groupId>
-            <artifactId>spring-webmvc</artifactId>
-            <version>${org.springframework.version}</version>
+            <groupId>org.glassfish.web</groupId>
+            <artifactId>jakarta.servlet.jsp.jstl</artifactId>
         </dependency>
         <dependency>
-            <groupId>taglibs</groupId>
-            <artifactId>standard</artifactId>
-            <version>1.1.2</version>
-        </dependency>
+            <groupId>org.apache.tomcat.embed</groupId>
+            <artifactId>tomcat-embed-jasper</artifactId>
+        </dependency>        
         <dependency>
-            <groupId>javax.servlet</groupId>
-            <artifactId>jstl</artifactId>
-            <version>1.2</version>
-        </dependency>
+            <groupId>com.fasterxml.jackson.core</groupId>
+            <artifactId>jackson-databind</artifactId>
+        </dependency>   
         <dependency>
-            <groupId>com.github.ryanbrainard</groupId>
-            <artifactId>richsobjects-core</artifactId>
-            <version>${com.github.ryanbrainard.richsobjects.version}</version>
-        </dependency>
+            <groupId>com.force.api</groupId>
+            <artifactId>force-metadata-api</artifactId>
+            <version>61.1.0</version>
+        </dependency>          
         <dependency>
-            <groupId>com.github.ryanbrainard</groupId>
-            <artifactId>richsobjects-api-jersey-client</artifactId>
-            <version>${com.github.ryanbrainard.richsobjects.version}</version>
-        </dependency>
+            <groupId>com.force.api</groupId>
+            <artifactId>force-partner-api</artifactId>
+            <version>61.1.0</version>
+        </dependency>      
         <dependency>
-            <groupId>com.github.ryanbrainard</groupId>
-            <artifactId>richsobjects-cache-memcached</artifactId>
-            <version>${com.github.ryanbrainard.richsobjects.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.force.sdk</groupId>
-            <artifactId>force-oauth</artifactId>
-            <version>${com.force.sdk.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.force.sdk</groupId>
-            <artifactId>force-springsecurity</artifactId>
-            <version>${com.force.sdk.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>jcl-over-slf4j</artifactId>
-            <version>${org.slf4j.version}</version>
-            <scope>runtime</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>slf4j-simple</artifactId>
-            <version>${org.slf4j.version}</version>
-            <scope>runtime</scope>
-        </dependency>
-        <dependency>
-            <groupId>com.github.jsimone</groupId>
-            <artifactId>webapp-runner</artifactId>
-            <version>8.0.30.2</version>
-            <scope>provided</scope>
+            <groupId>com.force.api</groupId>
+            <artifactId>force-wsc</artifactId>
+            <version>61.1.0</version>
         </dependency>
         <dependency>
         	<groupId>org.eclipse.mylyn.github</groupId>
